Renaming from spree to solidus
